% Generated by roxygen2: do not edit by hand
\name{convert_with_mapping}
\alias{convert_with_mapping}
\title{Convert with interactive column mapping (wrapper for Python CLI)}
\usage{
convert_with_mapping(input_file)
}
\arguments{
\item{input_file}{Character. Path to input file}
}
\value{
Information message about using Python CLI
}
\description{
This function provides information about using the Python CLI for
interactive conversion. R doesn't support interactive stdin well,
so use the Python CLI directly.
}
\examples{
\dontrun{
convert_with_mapping("data.tsv")
}
}
